# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real Estate API",
        "description": "REST API for the Real Estate platform. All endpoints are prefixed with /api/v1.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "List properties",
                "description": "Returns a paginated list of properties with optional filters.",
                "operationId": "13fea573f37f88995828f064090ca444",
                "parameters": [
                    {
                        "name": "listing_type",
                        "in": "query",
                        "description": "Filter by listing type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sale",
                                "rent"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by property type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "house",
                                "apartment",
                                "commercial",
                                "land"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (default: available)",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "sold",
                                "rented",
                                "reserved"
                            ]
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_area",
                        "in": "query",
                        "description": "Minimum square meters",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_area",
                        "in": "query",
                        "description": "Maximum square meters",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "description": "Minimum number of bedrooms",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathrooms",
                        "in": "query",
                        "description": "Minimum number of bathrooms",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "energy_class",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "A"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title, description, address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated property list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get a property",
                "operationId": "ed6a5eab533a0c91ff474042dd0183dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Property"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/properties/featured": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get featured properties",
                "operationId": "0b6254e3693c2e2acda6030e446cb507",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Featured properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/properties/search": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Advanced property search",
                "description": "Accepts the same filters as GET /properties.",
                "operationId": "3c9032684eef8cd4d17a5e43ff68070a",
                "parameters": [
                    {
                        "name": "listing_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sale",
                                "rent"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "house",
                                "apartment",
                                "commercial",
                                "land"
                            ]
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "min_area",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_area",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bathrooms",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "energy_class",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/properties/{id}/similar": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get similar properties",
                "operationId": "5f49757aace0a3ba2911dad762f03f42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Similar properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Seo": {
                "properties": {
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Property": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Modern apartment in Lisbon"
                    },
                    "slug": {
                        "type": "string",
                        "example": "modern-apartment-in-lisbon"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "type": {
                        "type": "string",
                        "example": "apartment",
                        "enum": [
                            "house",
                            "apartment",
                            "commercial",
                            "land"
                        ]
                    },
                    "listing_type": {
                        "type": "string",
                        "example": "sale",
                        "enum": [
                            "sale",
                            "rent"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "example": "available",
                        "enum": [
                            "available",
                            "sold",
                            "rented",
                            "reserved"
                        ]
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 250000
                    },
                    "price_formatted": {
                        "type": "string",
                        "example": "€250,000"
                    },
                    "bedrooms": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "bathrooms": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "square_meters": {
                        "type": "number",
                        "format": "float",
                        "example": 120.5,
                        "nullable": true
                    },
                    "year_built": {
                        "type": "integer",
                        "example": 2010,
                        "nullable": true
                    },
                    "energy_class": {
                        "type": "string",
                        "example": "A",
                        "nullable": true
                    },
                    "garage": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "postal_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "featured_image": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uri"
                        }
                    },
                    "extra_details": {
                        "type": "object"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "city": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/City"
                            }
                        ],
                        "nullable": true
                    },
                    "district": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/District"
                            }
                        ],
                        "nullable": true
                    },
                    "agent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Agent"
                            }
                        ],
                        "nullable": true
                    },
                    "seo": {
                        "$ref": "#/components/schemas/Seo"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}